# Fix bug #261: PDF indexing performance investigation and optimization