--- a/pyproject.toml
+++ b/pyproject.toml
@@ -51,6 +51,7 @@ dependencies = [
     "pydantic>=2.11.5",
     "pypdf>=5.6.0",
     "cryptography>=45.0.4",
+    "reportlab>=4.4.1",
 ]
 
 [project.scripts]