… PDF_INVESTIGATION_SUMMARY.md
… PDF_PERFORMANCE_ANALYSIS.md
… src/oboyu/crawler/services/enhanced_pdf_processor.py
… tests/fixtures/pdf/create_test_pdfs.py
… tests/fixtures/pdf/performance_tests/large_reliable.pdf
… tests/fixtures/pdf/performance_tests/medium_reliable.pdf
… tests/fixtures/pdf/performance_tests/small_reliable.pdf
… tests/fixtures/pdf/performance_tests/xl_reliable.pdf
… tests/fixtures/pdf/reliable_pdf_generator.py
… tests/performance/test_pdf_performance.py
… uv.lock